앱, 라우트, js, pug, css 수정함
