--- team_fridge/views/fridge/shopmemo 수정전.pug
@@ -1,25 +1,0 @@
-extends ../layout_home
-
-block content
-    link(rel="stylesheet", href="/css/shopmemo.css")
-    script(src="/js/shopmemo.js")
- 
-    div.body
-        section.box 
-            p 장보기 목록을 추가하세요!
-            section.input 
-                div.name
-                    input.name(placeholder="살것")
-                div.amount
-                    button.plus +
-                    input.amount(type="number", min="1", placeholder="수량")
-                    button.minus -
-            button.addmemo 추가
-            //- 장바구니 리스트
-            section.list 
-                ul 
-            section.stiky 
-                button.delete 전체삭제
-                button.save 저장
-       
-
